src/app/backend/requirements.txt
langchain==0.3.25
langchain-community==0.3.21
langchain-core==0.3.65
langchain-huggingface==0.3.0
langchain-ollama==0.3.3
python-dotenv==1.0.0
psycopg2-binary==2.9.10
fastapi==0.115.12
sqlalchemy==2.0.41
pydantic==2.11.5
markdown==3.7
streamlit==1.29.0
requests==2.32.4
bs4==0.0.2
sentence_transformers==4.1.0
faiss-cpu==1.11.0 
apscheduler==3.11.0
uvicorn==0.34.3